+++ src/assets/images/image.ts
@@ -58,4 +58,5 @@ export default {
   refresh: require('./refresh.png'),
   missionPointImage: require('./missionPointImage.png'),
   droneList: require('./droneList.png'),
+  emptyGuru: require('./emptyGuru.png'),
 };