+++ src/screens/GuruDetailScreen/SectionFooter.tsx
@@ -1,20 +1,42 @@
-import {Image, StyleSheet, TouchableOpacity, View} from 'react-native';
-import React from 'react';
+import {
+  Dimensions,
+  Image,
+  StyleSheet,
+  TouchableOpacity,
+  View,
+} from 'react-native';
+import React, {useEffect} from 'react';
 import Text from '../../components/Text';
 import {colors, font, icons} from '../../assets';
 import {numberWithCommas} from '../../function/utility';
 import {SheetManager} from 'react-native-actions-sheet';
+import RenderHTML from 'react-native-render-html';
+import {useDebounce} from '../../hooks/useDebounce';
+import {GuruKaset} from '../../datasource/GuruDatasource';
 
 interface Props {
   loveCount: number;
   commentCount: number;
+  description: string;
+  favorite?: boolean;
+  guruId: string;
+  isLoved?: boolean;
+  setIsLoved: React.Dispatch<React.SetStateAction<boolean>>;
 }
-export default function SectionFooter({loveCount, commentCount}: Props) {
-  const isOdd = Math.round(Math.random() * 10) % 2 === 0;
-  const [isLoved, setIsLoved] = React.useState(isOdd);
+export default function SectionFooter({
+  loveCount = 0,
+  commentCount,
+  description,
+  isLoved = false,
+  setIsLoved,
+  guruId,
+}: Props) {
+  const [isFirst, setIsFirst] = React.useState(true);
   const onPressLove = () => {
     setIsLoved(prev => !prev);
+    setIsFirst(false);
   };
+  const debounceValue = useDebounce(isLoved, 1000);
   const onOpenComment = async () => {
     await SheetManager.show('commentSheet', {
       payload: {
@@ -22,36 +44,42 @@ export default function SectionFooter({loveCount, commentCount}: Props) {
       },
     });
   };
-  const mockText = `
-  “ยาเหลือง” ใช้แล้วติดถังพ่น!! นักบินโดรนสามารถใช้ยาอะไร ทดแทนได้บ้าง!! วันนี้กูรูเกษตรมีคำตอบ ?
+  useEffect(() => {
+    const updateFavorite = async () => {
+      try {
+        await GuruKaset.updateFavoriteGuru({
+          guruId,
+        });
+        console.log('success :>>  liked');
+      } catch (error) {
+        console.log('error_like', error);
+      }
+    };
+    if (!isFirst) {
+      updateFavorite();
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [debounceValue, guruId]);
 
-  “ยาเหลือง” คือ สารเพนดิเมทาลิน 33% EC เป็นสารกำจัดวัชพืช ใช้ในช่วงก่อนวัชพืชงอก และวัชพืชกำลังงอก(pre-emergene & early post-emergene) กำจัดวัชพืชได้ทั้งใบแคบ และวัชพืชใบกว้าง ในนาหว่านข้าวแห้ง และพืชไร่ เช่น อ้อย มันสำปะหลัง ตัวยาจะ
-  ออกฤทธิ์ดูดซึมเข้าทางปลายรากฝอยหรือราก
-  แขนงและทางปลายยอดอ่อนของวัชพืชที่งอก
-  จากเมล็ด สารเพนดิเมทาลิน ลักษณะเนื้อยาจะมีสีเหลือง
-  เข้มถึงแม้จะนำไปเจือจางก่อนนำไปใช้งานก็ยังคงมีสีเหลือง ทำให้ถังเครื่องพ่นยา หรือถังโดรน
-  พ่นยามักจะติดสีเหลืองนี้ไปด้วย ล้างออกยาก หากล้างไม่ดี ไม่สะอาด อาจส่งผลเสียต่อพืชได้
-  กูรูเกษตร ขอแนะนำยาที่สามารถใช้ทดแทน “ยาเหลือง” ดังนี้
-  `;
   return (
     <View style={styles.container}>
       <View style={styles.sectionTop}>
         <TouchableOpacity style={styles.row} onPress={onPressLove}>
           <Image
-            source={isLoved ? icons.loveIcon : icons.loveFill}
+            source={isLoved ? icons.loveFill : icons.loveIcon}
             style={styles.icon}
           />
           <Text style={styles.textMedium}>
-            ถูกใจ
+            ถูกใจได้เลย
             {loveCount > 0 && (
               <Text style={styles.textCount}>{` ${numberWithCommas(
-                loveCount.toString(),
+                loveCount?.toString(),
                 true,
               )}`}</Text>
             )}
           </Text>
         </TouchableOpacity>
-        <TouchableOpacity style={styles.row} onPress={onOpenComment}>
+        {/* <TouchableOpacity style={styles.row} onPress={onOpenComment}>
           <Image source={icons.commentIcon} style={styles.icon} />
           <Text style={styles.textMedium}>
             ความคิดเห็น
@@ -62,10 +90,26 @@ export default function SectionFooter({loveCount, commentCount}: Props) {
               )}`}</Text>
             )}
           </Text>
-        </TouchableOpacity>
+        </TouchableOpacity> */}
       </View>
       <View style={styles.content}>
-        <Text style={styles.textContent}>{mockText}</Text>
+        {/* <Text style={styles.textContent}>{}</Text> */}
+        <RenderHTML
+          source={{html: description}}
+          defaultTextProps={{
+            allowFontScaling: false,
+            style: styles.textContent,
+          }}
+          contentWidth={Dimensions.get('window').width - 64}
+          tagsStyles={{
+            body: {
+              fontSize: 16,
+              color: colors.fontBlack,
+              fontFamily: font.light,
+            },
+          }}
+          systemFonts={[font.light, font.semiBold, font.medium, font.bold]}
+        />
       </View>
     </View>
   );