+++ src/screens/GuruScreen/index.tsx
@@ -1,10 +1,12 @@
 import {
   Animated,
+  Dimensions,
   FlatList,
   Image,
   NativeScrollEvent,
   NativeSyntheticEvent,
   Pressable,
+  RefreshControl,
   SafeAreaView,
   ScrollView,
   StyleSheet,
@@ -18,37 +20,75 @@ import {colors, font, icons} from '../../assets';
 import GuruTapNavigator from './GuruTapNavigator';
 import ItemContent from './ItemContent';
 import {GuruKaset} from '../../datasource/GuruDatasource';
+import {useIsFocused} from '@react-navigation/native';
+import SkeletonPlaceholder from 'react-native-skeleton-placeholder';
+import EmptyGuru from '../../components/EmptyGuru';
 
 interface Props {
   navigation: any;
 }
 const mockData = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
-const initialListsHeader = [
-  {
-    title: 'ทั้งหมด',
-    value: 'all',
-  },
-  {
-    title: 'โรคพืช',
-    value: 'plantDisease',
-  },
-  {
-    title: 'การใช้ยา',
-    value: 'useMedicine',
-  },
-  {
-    title: 'การใช้ปุ๋ย',
-    value: 'useFertilizer',
-  },
-  {
-    title: 'การใช้สารเคมี',
-    value: 'useChemical',
-  },
-];
+// const initialListsHeader = [
+//   {
+//     title: 'ทั้งหมด',
+//     value: 'all',
+//   },
+//   {
+//     title: 'โรคพืช',
+//     value: 'plantDisease',
+//   },
+//   {
+//     title: 'การใช้ยา',
+//     value: 'useMedicine',
+//   },
+//   {
+//     title: 'การใช้ปุ๋ย',
+//     value: 'useFertilizer',
+//   },
+//   {
+//     title: 'การใช้สารเคมี',
+//     value: 'useChemical',
+//   },
+// ];
+const limit = 10;
+export interface GuruData {
+  _id: string;
+  type: string;
+  name: string;
+  view: number;
+  like: number;
+  commentCount: number;
+  read: number;
+  application: string;
+  status: string;
+  grouping: string;
+  startDate: string | null;
+  createdAt: string;
+  updatedAt: string;
+  __v: number;
+  description: string;
+  image: string;
+  startDateCronJob: string | null;
+  favorite: boolean;
+  groupName: string;
+}
+interface GuruKasetData {
+  data: GuruData[];
+  count: 0;
+}
+const initialPage = 1;
 export default function GuruScreen({navigation}: Props) {
-  const [listsHeader, setListsHeader] = React.useState(initialListsHeader);
+  const [listsHeader, setListsHeader] = React.useState([]);
+  const isFocused = useIsFocused();
+  const [loading, setLoading] = React.useState(false);
+  const [page, setPage] = React.useState(initialPage);
+  const [refreshing, setRefreshing] = React.useState(false);
   const [currentTabHeader, setCurrentTabHeader] = React.useState('all');
   const [currentTab, setCurrentTab] = React.useState('all');
+  const [guruKasetData, setGuruKasetData] = React.useState<GuruKasetData>({
+    data: [],
+    count: 0,
+  });
 
   const onPressBack = () => {
     navigation.goBack();
@@ -80,13 +120,75 @@ export default function GuruScreen({navigation}: Props) {
   useEffect(() => {
     const getInitialData = async () => {
       try {
-        const getListGroupGuru = await GuruKaset.getGroupGuru();
+        const resultGroup = await GuruKaset.getGroupGuru({
+          page: 1,
+          take: 20,
+        });
+        const formatGroupGuru = (resultGroup.data || []).map((item: any) => {
+          return {
+            title: item.groupName,
+            value: item._id,
+          };
+        });
+
+        setListsHeader(formatGroupGuru);
       } catch (e) {
         console.log(e);
       }
     };
+    getInitialData();
   }, []);
 
+  const getListGuru = async () => {
+    try {
+      setLoading(true);
+      const result = await GuruKaset.getAllGuru({
+        page: 1,
+        limit: limit,
+        groupId: currentTabHeader === 'all' ? undefined : currentTabHeader,
+      });
+      setGuruKasetData(result);
+      setLoading(false);
+    } catch (e) {
+      console.log(e);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  useEffect(() => {
+    getListGuru();
+  }, [currentTabHeader, isFocused]);
+
+  const onRefresh = async () => {
+    try {
+      setRefreshing(true);
+      getListGuru();
+      setRefreshing(false);
+    } catch (e) {
+      console.log(e);
+    }
+  };
+  const onLoadMore = async () => {
+    if (guruKasetData.data.length >= guruKasetData.count) {
+      console.log('no more data');
+      return;
+    }
+    try {
+      const result = await GuruKaset.getAllGuru({
+        page: page + 1,
+        limit: limit,
+        groupId: currentTabHeader === 'all' ? undefined : currentTabHeader,
+      });
+      setGuruKasetData({
+        data: [...guruKasetData.data, ...result.data],
+        count: result.count,
+      });
+      setPage(page + 1);
+    } catch (e) {
+      console.log(e);
+    }
+  };
   return (
     <SafeAreaView style={{flex: 1, backgroundColor: colors.white}}>
       <View style={styles.container}>
@@ -102,27 +204,61 @@ export default function GuruScreen({navigation}: Props) {
               }}
             />
           </TouchableOpacity>
-          <GuruTapNavigator
+          <View
+            style={{
+              paddingTop: 14,
+              flexDirection: 'row',
+              width: '100%',
+              marginLeft: Dimensions.get('window').width / 2 - 140,
+            }}>
+            <Text
+              style={{
+                fontFamily: font.bold,
+                fontSize: 20,
+              }}>
+              บทความทั้งหมด
+            </Text>
+          </View>
+          {/* <GuruTapNavigator
             onChange={(value: string) => {
               setCurrentTab(value);
             }}
             value={currentTab}
-          />
+          /> */}
         </View>
         <MemoHeaderFlatList
           lists={listsHeader}
           value={currentTabHeader}
           onChange={(value: string) => {
             setCurrentTabHeader(value);
+            setPage(initialPage);
           }}
         />
         <FlatList
+          refreshControl={
+            <RefreshControl refreshing={refreshing} onRefresh={onRefresh} />
+          }
+          scrollEnabled={guruKasetData?.data?.length > 0}
+          ListEmptyComponent={<EmptyGuru />}
           scrollEventThrottle={64}
+          onEndReached={onLoadMore}
           onScroll={handleScroll}
-          data={mockData}
+          data={
+            loading
+              ? ([
+                  {
+                    _id: '1',
+                  },
+                  {
+                    _id: '2',
+                  },
+                ] as GuruData[])
+              : guruKasetData.data || []
+          }
           contentContainerStyle={{
             paddingHorizontal: 16,
           }}
+          keyExtractor={(item, index) => `${item?._id}-${index}`}
           // ListHeaderComponent={() => {
           //   return (
           //     <Animated.View
@@ -131,8 +267,53 @@ export default function GuruScreen({navigation}: Props) {
           //       }}></Animated.View>
           //   );
           // }}
-          renderItem={() => {
-            return <ItemContent navigation={navigation} />;
+          renderItem={({item, index}) => {
+            if (loading) {
+              return (
+                <View style={{height: 450}}>
+                  <SkeletonPlaceholder
+                    backgroundColor={colors.skeleton}
+                    speed={2000}
+                    borderRadius={10}>
+                    <>
+                      <SkeletonPlaceholder.Item
+                        height={350}
+                        borderRadius={10}
+                        marginBottom={8}
+                      />
+                      <SkeletonPlaceholder.Item
+                        height={14}
+                        borderRadius={4}
+                        marginBottom={4}
+                      />
+                      <SkeletonPlaceholder.Item height={14} borderRadius={4} />
+                      <SkeletonPlaceholder.Item
+                        marginTop={8}
+                        flexDirection="row"
+                        alignItems="center"
+                        justifyContent="space-between">
+                        <View
+                          style={{
+                            width: 30,
+                            height: 30,
+                            borderRadius: 20,
+                          }}
+                        />
+                        <View
+                          style={{
+                            width: 100,
+                            height: 20,
+                          }}
+                        />
+                      </SkeletonPlaceholder.Item>
+                    </>
+                  </SkeletonPlaceholder>
+                </View>
+              );
+            }
+            return (
+              <ItemContent navigation={navigation} item={item} key={index} />
+            );
           }}
         />
       </View>
@@ -149,6 +330,7 @@ const styles = StyleSheet.create({
     alignItems: 'flex-start',
     borderBottomWidth: 1,
     borderBottomColor: colors.disable,
+    height: 50,
   },
   item: {
     width: '100%',
@@ -164,52 +346,56 @@ interface HeaderFlatListProps {
   onChange?: (value: string) => void;
 }
 const HeaderFlatList = ({lists = [], onChange, value}: HeaderFlatListProps) => {
-  const scaleAnim = useRef(new Animated.Value(1)).current;
+  // const scaleAnim = useRef(new Animated.Value(1)).current;
   const handlePress = (itemValue: string) => {
+    if (itemValue === value) {
+      onChange && onChange('all');
+
+      return;
+    }
     if (onChange) {
       onChange(itemValue);
     }
   };
-  useEffect(() => {
-    Animated.sequence([
-      Animated.timing(scaleAnim, {
-        toValue: 1.1,
-        duration: 100,
-        useNativeDriver: true,
-      }),
-      Animated.timing(scaleAnim, {
-        toValue: 1,
-        duration: 100,
-        useNativeDriver: true,
-      }),
-    ]).start();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [value]);
+  // useEffect(() => {
+  //   Animated.sequence([
+  //     Animated.timing(scaleAnim, {
+  //       toValue: 1.1,
+  //       duration: 100,
+  //       useNativeDriver: true,
+  //     }),
+  //     Animated.timing(scaleAnim, {
+  //       toValue: 1,
+  //       duration: 100,
+  //       useNativeDriver: true,
+  //     }),
+  //   ]).start();
+  //   // eslint-disable-next-line react-hooks/exhaustive-deps
+  // }, [value]);
   return (
-    <ScrollView
-      horizontal
-      showsHorizontalScrollIndicator={false}
-      style={{
-        paddingVertical: 8,
-        paddingHorizontal: 16,
-        height: 'auto',
-        marginBottom: 16,
-      }}>
-      {lists.map((item, index) => {
-        const isActive = value === item.value;
-        return (
-          <Animated.View
-            key={index}
-            style={{
-              transform: [{scale: isActive ? scaleAnim : 1}],
-              height: 'auto',
-            }}>
+    <View>
+      <ScrollView
+        horizontal
+        showsHorizontalScrollIndicator={false}
+        style={{
+          paddingVertical: 8,
+          paddingHorizontal: 16,
+          height: 'auto',
+          marginBottom: 16,
+        }}>
+        {lists.map((item, index) => {
+          const isActive = value === item.value;
+
+          return (
+            // <Animated.View
+            //   key={index}
+            //   style={{
+            //     transform: [{scale: isActive ? scaleAnim : 1}],
+            //     height: 'auto',
+            //   }}>
             <Pressable
               key={index}
               onPress={() => {
-                if (isActive) {
-                  return;
-                }
                 handlePress(item.value);
               }}
               style={{
@@ -232,10 +418,11 @@ const HeaderFlatList = ({lists = [], onChange, value}: HeaderFlatListProps) => {
                 {item.title}
               </Text>
             </Pressable>
-          </Animated.View>
-        );
-      })}
-    </ScrollView>
+            // </Animated.View>
+          );
+        })}
+      </ScrollView>
+    </View>
   );
 };
 const MemoHeaderFlatList = React.memo(HeaderFlatList);