+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 
+## Feature [1.10.5] 2023-12-01
+- guru kaset screen
 ## Fix [1.10.4] 2023-11-29
 - Pending point bug ui
 ## Enhance [1.10.2] 2023-11-27